requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Flask>=1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy>=1.20.1
pandas>=1.2.2
python-dateutil==2.8.1
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
scikit-learn>=0.24.1
scipy==1.6.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2
threadpoolctl==2.1.0
urllib3==1.26.3
webdriver-manager==3.3.0
Werkzeug==1.0.1
